Move orphans to type class declaration site
