+++ index-server/ergvein-index-server.cabal
@@ -30,7 +30,6 @@ library
     Ergvein.Index.Server.DB.Schema.Indexer
     Ergvein.Index.Server.DB.Serialize
     Ergvein.Index.Server.DB.Serialize.Class
-    Ergvein.Index.Server.DB.Serialize.Tx
     Ergvein.Index.Server.DB.Utils
     Ergvein.Index.Server.Dependencies
     Ergvein.Index.Server.Environment
