+++ index-server/src/Ergvein/Index/Server/DB/Serialize.hs
@@ -16,7 +16,6 @@ import Data.ByteString.Builder as BB
 import Data.Foldable
 import Data.Word
 
-import Ergvein.Index.Server.DB.Serialize.Tx ()
 import Ergvein.Index.Server.BlockchainScanning.Types
 import Ergvein.Index.Server.PeerDiscovery.Types
 import Ergvein.Index.Server.DB.Schema.Filters
